Improve inequality performance on sorted columns
When a query predicate is specified on a sorted column, we gather the
matching dictionaryIds and convert them to a range list. For inequality
predicates, such as NEQ (not equals) or NOT_IN, when querying a high
cardinality column, it means that we have to hold on to a large list
of ranges during query processing.

This rewritten algorithm for computing the list of ranges merges
adjacent docId ranges and builds its list of ranges from the list of
excluded dictionaryIds, thus minimizing the number of ranges held
resident during query processing and lowering the amount of memory
used during effective range calculation.